Optionalize target support
Issue #3
